backend/pyproject.toml
[project]
name = "llmops-product-backend"
version = "0.1.0"
description = "Backend RAG pour l'assistant de conformité réglementaire"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-community>=0.0.10",
    "sentence-transformers>=2.2.2",
    "openai>=1.6.0",
    "python-docx>=1.1.0",
    "pypdf>=3.17.0",
    "tiktoken>=0.5.1",
    "python-dotenv>=1.0.0",
    "alembic>=1.12.1",
    "sse-starlette>=1.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

